fix: Critical production fixes for RTX 5090 support
- Fix server_args.py syntax error (removed duplicate import os)
- Fix C++ linkage: removed extern "C" with std::vector (ill-formed)
- Add common header for GroupedDesc struct shared between .cpp and .cu
- Add missing ATen/cuda/CUDAContext.h include for getCurrentCUDAStream
- Create kernels/__init__.py to mark as Python package for imports
- Clean up warning messages with arrow notation

These fixes resolve build and runtime errors for MXFP4 weight-only kernels on RTX 5090.
